moved common logic to utils file
